alex-companion/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
edge-tts
aiofiles
